% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors}
\alias{colors}
\title{Example Data Frame of Color Vectors}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 300 columns.
}
\source{
Schmitz, D., & Schneider, V. (2022). gdsm: General functions for Distributional SeMantics. R package version 0.1. URL: https://github.com/dosc91/gdsm
}
\usage{
data(colors)
}
\description{
This data frame contains color vectors of 300 dimensions. It is mainly used for exercises.
}
\keyword{datasets}
