% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_fast.R
\docType{data}
\name{colors_fast}
\alias{colors_fast}
\title{Example Data Frame of Color Vectors created with fastText}
\format{
An object of class \code{data.frame} with 10 rows and 300 columns.
}
\source{
Schmitz, D., & Schneider, V. (2022). gdsm: General functions for Distributional SeMantics. R package version 0.1. URL: https://github.com/dosc91/gdsm
}
\usage{
data(colors_fast)
}
\description{
This data frame contains color vectors. It is mainly used for exercises.
}
\keyword{datasets}
