% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosim.R
\name{cosim}
\alias{cosim}
\title{Cosine Similarity}
\usage{
cosim(v1, v2, data = NULL)
}
\arguments{
\item{v1}{Either an individual vector or a row name of \code{data}.}

\item{v2}{Either an individual vector or a row name of \code{data}.}

\item{data}{A data frame object or matrix.}
}
\description{
\code{cosim} computes the cosine similarity of two vectors.
Specify either two individual vectors or a data frame/matrix and two pertinent row names as input.
}
\examples{

### two individual vectors

var01 <- runif(n = 50, min = 1, max = 10)
var02 <- runif(n = 50, min = 1, max = 10)

cosim(var01, var02)


### data frame

data("gdsm_df")

cosim("var03", "var04", gdsm_df)


### matrix

data("gdsm_mat")

cosim("var05", "var06", gdsm_mat)

}
\author{
D. Schmitz
}
