% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid.R
\name{euclid}
\alias{euclid}
\title{Euclidian Distance}
\usage{
euclid(v1, v2, data = NULL)
}
\arguments{
\item{v1}{Either an individual vector or a row name of \code{data}.}

\item{v2}{Either an individual vector or a row name of \code{data}.}

\item{data}{A data frame object or matrix.}
}
\value{
Nothing.
}
\description{
\code{euclid} computes the Euclidian distance between two vectors.
Specify either two individual vectors or a data frame/matrix and two pertinent row names as input.
}
\examples{

### two individual vectors

var01 <- runif(n = 50, min = 1, max = 10)
var02 <- runif(n = 50, min = 1, max = 10)

euclid(var01, var02)


### data frame

data("gdsm_df")

euclid("var03", "var04", gdsm_df)


### matrix

data("gdsm_mat")

euclid("var05", "var06", gdsm_mat)

}
\author{
D. Schmitz
}
