% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsm_df.R
\docType{data}
\name{gdsm_df}
\alias{gdsm_df}
\title{Example Data Frame of Vectors}
\format{
An object of class \code{data.frame} with 20 rows and 50 columns.
}
\source{
Schmitz, D., & Schneider, V. (2022). gdsm: General functions for Distributional SeMantics. R package version 0.1. URL: https://github.com/dosc91/gdsm
}
\usage{
data(gdsm_df)
}
\description{
This data frame contains random vectors of 50 dimensions. It is mainly used for illustration purposes.
}
\keyword{datasets}
