% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship_fast.R
\docType{data}
\name{kinship_fast}
\alias{kinship_fast}
\title{Example Data Frame of Kinship Vectors created with fastText}
\format{
An object of class \code{data.frame} with 10 rows and 300 columns.
}
\source{
Schmitz, D., & Schneider, V. (2022). gdsm: General functions for Distributional SeMantics. R package version 0.1. URL: https://github.com/dosc91/gdsm
}
\usage{
data(kinship_fast)
}
\description{
This data frame contains kinship vectors. It is mainly used for exercises.
}
\keyword{datasets}
