% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_density.R
\name{n_density}
\alias{n_density}
\title{Calculate Neighborhood Density}
\usage{
n_density(vec, neighbors, n = 8)
}
\arguments{
\item{vec}{The vector for which the neighbor density is to be calculated. This is either a vector saved as object or a row of the \code{neighbors} data frame or matrix. If a row of a data frame or matrix is specified, this row is excluded as competitor for the nearest neighbor (as a row will be closest to itself).}

\item{neighbors}{The potential neighbors, usually given as data frame or matrix.}

\item{n}{The number of neighbors wanted as output. Defaults to \code{8}.}
}
\description{
\code{n_density} takes the \code{n} most strongly correlated neighbors of a target vector and their correlation coefficients to calculate a mean correlation
coefficient. The higher the value of this coefficient, the denser the neighborhood.
}
\examples{

### a vector saved as object & a matrix of neighbors

vector <- runif(50, 0, 10)
data("gdsm_mat")

n_density(vec = vector, neighbors = gdsm_mat)

n_density(vec = vector, neighbors = gdsm_mat, n = 4)


### a vector specified by its name & its matrix of neighbors

data("gdsm_mat")

n_density(vec = "var12", neighbors = gdsm_mat)

n_density(vec = "var12", neighbors = gdsm_mat, n = 4)

}
\author{
D. Schmitz
}
