% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_exercises.R
\name{open_exercises}
\alias{open_exercises}
\title{Open Exercises}
\usage{
open_exercises(session)
}
\arguments{
\item{session_number}{The number of the session.}
}
\description{
Calling \code{open_exercises} with a session number as argument will open the exercise html for that session.
The html will be opened in your default browser. This function is inspired by Schmitz & Esser (see references).
}
\examples{
open_exercises(02)

}
\references{
Schmitz, D., & Esser, J. (2021). SfL: Statistics for Linguistics. R package version 0.3. URL: https://github.com/dosc91/SfL
}
\author{
D. Schmitz
}
