% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_slides.R
\name{open_slides}
\alias{open_slides}
\title{Open Slides}
\usage{
open_slides(session)
}
\arguments{
\item{session_number}{The number of the session.}
}
\description{
Calling \code{open_slides} with a session number as argument will open the slides for that session.
The slides will be opened in your default browser. This function is inspired by Schmitz & Esser (see references).
}
\examples{
open_slides(02)

}
\references{
Schmitz, D., & Esser, J. (2021). SfL: Statistics for Linguistics. R package version 0.3. URL: https://github.com/dosc91/SfL
}
\author{
D. Schmitz
}
