% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_matrix.R
\name{random_matrix}
\alias{random_matrix}
\title{Create Random Matrix}
\usage{
random_matrix(dims = 50, rows = 25, df = FALSE)
}
\arguments{
\item{dims}{The number of dimensions (columns) of the random matrix. Defaults to \code{50}.}

\item{rows}{The number of rows of the random matrix. Defaults to \code{25}.}

\item{df}{Specify whether the matrix is to be created as data frame object. Defaults to \code{FALSE}.}
}
\description{
\code{random_matrix} creates a matrix of randomly generated vectors.
}
\examples{

### using default settings

matrix_a <- random_matrix()
dim(matrix_a)

### as data frame object

matrix_b <- random_matrix(df = TRUE)
dim(matrix_b)

### more dimensions

matrix_c <- random_matrix(dims = 100)
dim(matrix_c)

### more rows

matrix_d <- random_matrix(rows = 100)
dim(matrix_d)

}
\author{
D. Schmitz
}
